// src/app/interfaces/condicion5.ts
export interface Condicion5 {
    condicion5Aspecto1: string;
    condicion5Aspecto2: string;
    condicion5Aspecto3: string;
    condicion5Aspecto4: string;
    condicion5Aspecto5: string;
    condicion5Aspecto6: string;
    archivoAspecto1: string;
    archivoAspecto2: string;
    archivoAspecto3: string;
    archivoAspecto4: string;
    archivoAspecto5: string;
    archivoAspecto6: string;
    key$?: string;
}